requirements.txt
absl-py==2.1.0
astunparse==1.6.3
attrs==23.2.0
beautifulsoup4==4.12.3
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
filelock==3.13.1
flatbuffers==23.5.26
fsspec==2023.12.2
gast==0.5.4
google-api-core==2.16.2
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
googleapis-common-protos==1.62.0
grpcio==1.60.1
h11==0.14.0
h5py==3.10.0
httplib2==0.22.0
huggingface-hub==0.20.3
idna==3.6
joblib==1.3.2
keras==2.15.0
libclang==16.0.6
lxml==5.1.0
Markdown==3.5.2
MarkupSafe==2.1.5
ml-dtypes==0.2.0
nltk==3.8.1
numpy==1.26.3
oauthlib==3.2.2
opt-einsum==3.3.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.0
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.4
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.2
selenium==4.17.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tokenizers==0.15.1
tqdm==4.66.1
transformers==4.37.2
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.0
Werkzeug==3.0.1
wrapt==1.14.1
wsproto==1.2.0
